extract pool names according to storageSystem name
